+++ modules/apps/commerce/commerce-order-content-web/src/main/resources/META-INF/resources/js/multishipping/DeliveryGroupHeaderCell.tsx
