+++ modules/apps/commerce/commerce-order-content-web/test/components/DeliveryGroupModal/DeliveryGroupModal.spec.tsx
@@ -11,14 +11,14 @@ import {
 	ICountryAPIResponse,
 	IPostalAddress,
 	IPostalAddressAPIResponse,
-} from '../../../src/main/resources/META-INF/resources/js/multi_shipping/Types';
+} from '../../../src/main/resources/META-INF/resources/js/multishipping/Types';
 
 import '@testing-library/jest-dom/extend-expect';
 import {RenderResult, cleanup, render, waitFor} from '@testing-library/react';
 import React from 'react';
 import {act} from 'react-dom/test-utils';
 
-import DeliveryGroupModal from '../../../src/main/resources/META-INF/resources/js/multi_shipping/DeliveryGroupModal';
+import DeliveryGroupModal from '../../../src/main/resources/META-INF/resources/js/multishipping/DeliveryGroupModal';
 import {setFieldValue} from '../../utils/utils.spec';
 
 interface ILocators {
